refactor(order): forgot some modif
